+++ js/main.js
@@ -2,11 +2,10 @@ import {renderMarkers} from './map-set.js';
 import {setUserFormSubmit} from './advertisment-form.js';
 import {getData} from './api.js';
 import {getAdTemplate} from './draw-advertisment.js';
-import {typeCLick, roomsClick, guestsClick, priceClick, featuresClick,getFiltersData} from './map-filters.js';
-import {debounce} from './utils/debounce.js';
+import {onFiltersFormChange,getFiltersData} from './map-filters.js';
+import {debounce} from './util.js';
 
 const RERENDER_DELAY = 500;
-
 
 getData((advertisments) => {
   renderMarkers(advertisments,getAdTemplate);
@@ -14,13 +13,8 @@ getData((advertisments) => {
     const filteredAdvertisments  = getFiltersData(advertisments);
     renderMarkers(filteredAdvertisments,getAdTemplate);
   };
-  typeCLick(debounce(() =>  onFilter()),RERENDER_DELAY);
-  guestsClick(debounce(() => onFilter()),RERENDER_DELAY);
-  roomsClick(debounce(() => onFilter()),RERENDER_DELAY);
-  priceClick(debounce(() => onFilter()),RERENDER_DELAY);
-  featuresClick(debounce(() => onFilter()),RERENDER_DELAY);
+  onFiltersFormChange(debounce(() => onFilter()),RERENDER_DELAY);
 });
-
 
 setUserFormSubmit();
